Create account

+++ frontend/src/views/Register.vue
@@ -19,8 +19,19 @@
                   <div
                     class="card-body mx-3"
                     style="display: flex; flex-direction: column; text-align: left">
-                    <h5 class="card-title">Login</h5>
-                    <form @submit.prevent="login">
+                    <h5 class="card-title">Register</h5>
+                    <form @submit.prevent="register">
+                      <div class="mb-3">
+                        <label class="form-label" for="email">Name</label>
+                        <input
+                          type="text"
+                          class="form-control"
+                          style="display: block"
+                          id="name"
+                          v-model="name"
+                          placeholder="Your name"
+                          required />
+                      </div>
                       <div class="mb-3">
                         <label class="form-label" for="email">Email address</label>
                         <input
@@ -37,34 +48,54 @@
                         <div class="input-group">
                           <input
                             :type="passwordVisible ? 'text' : 'password'"
-                            class="form-control"
-                            style="display: block"
+                            class="form-control d-block"
                             id="password"
                             v-model="password"
                             placeholder="Password"
                             required />
                           <div class="input-group-append">
                             <span class="input-group-text" @click="togglePassword">
-                              <i :class="passwordVisible ? 'fas fa-eye-slash' : 'fas fa-eye'" id="iconShowPass"></i>
+                              <i
+                                :class="passwordVisible ? 'fas fa-eye-slash' : 'fas fa-eye'"
+                                id="iconShowPass"></i>
                             </span>
                           </div>
                         </div>
                       </div>
-
+                      <div class="mb-3">
+                        <label class="form-label" for="passwordConfirm">Confirm Password</label>
+                        <div class="input-group">
+                          <input
+                            :type="confirmPasswordVisible ? 'text' : 'password'"
+                            class="form-control d-block"
+                            id="passwordConfirm"
+                            v-model="passwordConfirm"
+                            placeholder="Confirm Password"
+                            required />
+                          <div class="input-group-append">
+                            <span class="input-group-text" @click="toggleConfirmPassword">
+                              <i
+                                :class="confirmPasswordVisible ? 'fas fa-eye-slash' : 'fas fa-eye'"
+                                id="iconShowPass"></i>
+                            </span>
+                          </div>
+                        </div>
+                      </div>
                       <button
                         type="submit"
                         class="btn mb-3"
                         style="background-color: #22b3c1; color: aliceblue; width: 100%">
-                        Login
+                        Register
                       </button>
                     </form>
+                    <span class="text-center mb-2">Or</span>
                     <button
                       @click="loginWithGoogle"
                       class="btn mb-3 btn-primary"
                       style="width: 100%">
                       Login with Google
                     </button>
-                    <router-link class="text-button" to="/register">dont have any account?</router-link>
+                    <router-link class="text-button" to="/login">already have an account?</router-link>
                   </div>
                 </div>
               </div>
@@ -84,7 +115,7 @@ import PublicStyle from "@/components/public/PublicStyle.vue";
 import axios from "axios";
 
 export default {
-  name: "LoginPage",
+  name: "RegisterPage",
   components: {
     PublicScripts,
     PublicStyle,
@@ -94,28 +125,34 @@ export default {
       email: "",
       password: "",
       passwordVisible: false,
+      confirmPasswordVisible: false,
     };
   },
   methods: {
-    async login() {
-      try {
-        const response = await axios.post("http://localhost:5000/login", {
-          email: this.email,
-          password: this.password,
-        });
-        const token = response.data.token;
-        localStorage.setItem("token", token);
-
-        // Pindah ke halaman /home
-        this.$router.push("/home");
-
-        console.log("Login berhasil, token disimpan:", token);
-      } catch (error) {
-        if (error.response) {
-          alert(error.response.data.message); // Tampilkan pesan error
-        } else {
-          console.error(error);
+    async register() {
+      if (this.password == this.passwordConfirm) {
+        try {
+          const response = await axios.post("http://localhost:5000/register", {
+            name: this.name,
+            email: this.email,
+            password: this.password,
+          });
+          const token = response.data.token;
+          localStorage.setItem("token", token);
+  
+          // Pindah ke halaman /home
+          this.$router.push("/home");
+  
+          console.log("Registrasi berhasil, token disimpan:", token);
+        } catch (error) {
+          if (error.response) {
+            alert(error.response.data.message);
+          } else {
+            console.error(error);
+          }
         }
+      }else{
+        alert('Password tidak cocok');
       }
     },
     loginWithGoogle() {
@@ -123,7 +160,10 @@ export default {
     },
     togglePassword() {
       this.passwordVisible = !this.passwordVisible;
-    }
+    },
+    toggleConfirmPassword() {
+      this.confirmPasswordVisible = !this.confirmPasswordVisible;
+    },
   },
   created() {
     const urlParams = new URLSearchParams(window.location.search);
